Makefile: Only include cc_platform.mk when not using runc_nodmz
One of the reasons to have a build tag to disable runc-dmz is to get rid
of possible compile issues introduced by it (it needs to cross-compile a
C program).

We were not being strict enough to include this file only when needed.
This patch includes it only when not using runc_nodmz.

While we want to fix the compilation issues of runc-dmz in another
commit, let's make when using runc_nodmz we don't include more than
needed to compile.

Signed-off-by: Rodrigo Campos <rodrigoca@microsoft.com>