cc_platform.mk: only use when GOARCH is set explicitly
Current cc_platform.mk code assumes that if the value of GOARCH is the
same as `go env GOARCH` output, we have a native build. This assumption
is not correct when xx [1] is used to build runc (as it sets GOARCH etc
in ~/.config/go/env file).

Let's only include cc_platform when GOARCH is set from the environment
of the command line.

[1]: https://github.com/tonistiigi/xx

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>