init: do not close syncpipe before execve
If the execve fails, we end up returning errors through the sync pipe
and closing it causes other spurious errors to pop up. It seems
35aa63ea874f ("never send procError after the socket closed") tried to
fix this but the problem is that UnsafeCloseFrom() closes the file
descriptor without using Close() so the file descriptor still "looks
open".

There is no security risk in potentially leaking the pipe into the
container because it is a purely in-memory pipe that will have its write
end closed by the time the container could do anything with it.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>