libct/cg: retry opening a cgroup file
From time to time we see unexpected errors (ENODEV or ENOENT) trying to
write to a cgroup file. This is probably due to some races in the
kernel.

Let's implement a retry when we see ENOENT or ENODEV, but only if a
directory exists.

Note this can have a negative effect on performance when some cgroup
features are disabled.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>