Revert "increase memory.max in cgroups.bats"
This reverts commit 65a1074c75123afbe1d59e6ca2abc683e9d0d9dc.

We needed [1] because when we removed the bindfd logic in [2] we had not
yet moved the binary cloning logic to Go and thus it was necessary to
increase the memory limit in CI because the clone was happening after
joining the cgroup. However, [3] finally moved that code to Go and thus
the cloning is now done outside of the container's cgroup and thus is no
longer accounted as part of the container's memory usage at any point.

Now we can properly support running a simple container with lower memory
usage as we did before.

[1]: commit 65a1074c7512 ("increase memory.max in cgroups.bats")
[2]: commit b999376fb237 ("nsenter: cloned_binary: remove bindfd logic entirely")
[3]: commit 0e9a3358f848 ("nsexec: migrate memfd /proc/self/exe logic to Go code")

Signed-off-by: lfbzhm <lifubang@acmcoder.com>
[cyphar: fixed commit messages]
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>