tests: helpers: provide a "sane" run helper
The runc wrapper we have had for a long time is useful for debugging
errors (because bats's built-in "run" doesn't provide the output of the
program in case of an error). However, it might be necessary to run
other programs in a similar wrapper.

In addition, it might be needed to add timeouts or use the bats_pipe
feature (bats v1.10.0) with similar wrapping. Adding support for this in
the sane_run wrapper makes it a little easier to use these things when
writing tests.

This is somewhat adapted from umoci's sane_run helper.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>