[1.2] deps: update to github.com/cyphar/filepath-securejoin@v0.3.5
(This is a cherry-pick of 2f1b6626f38c63ee37930267caa3a9bf57a2ea79.)

This fixes a regression in use of securejoin.MkdirAll, where multiple
runc processes racing to create the same mountpoint in a shared rootfs
would result in spurious EEXIST errors. In particular, this regression
caused issues with BuildKit.

Fixes: dd827f7b715a ("utils: switch to securejoin.MkdirAllHandle")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>