libct/cg/sd: write rounded CPU quota to cgroupfs
When CPU quota is updated, the value is converted to CPUQuotaPerSecUSec
property for passing to systemd. The value can be rounded in the
following cases:
- The value is rounded up to the nearest 10ms.
- Depending on CPU period, the value may be rounded during division.
Because of this rounding, systemd and runc systemd driver may write
different values to cgroupfs. In order to avoid this inconsistency,
this fix makes systemd driver write the same rounded value to cgroupfs
by calculating the value from systemd properties.

Signed-off-by: Hironori Shiina <shiina.hironori@gmail.com>